transition GET /api/v1/get-upload-file
