+++ packages/server/src/index.ts
@@ -5,7 +5,6 @@ import cors from 'cors'
 import http from 'http'
 import * as fs from 'fs'
 import basicAuth from 'express-basic-auth'
-import contentDisposition from 'content-disposition'
 import { Server } from 'socket.io'
 import logger from './utils/logger'
 import { expressRequestLogger } from './utils/logger'
@@ -189,35 +188,6 @@ export class App {
             const fileStream = fs.createReadStream(filePath)
             fileStream.pipe(res)
         }
-
-        // stream uploaded image
-        this.app.get('/api/v1/get-upload-file', async (req: Request, res: Response) => {
-            try {
-                if (!req.query.chatflowId || !req.query.chatId || !req.query.fileName) {
-                    return res.status(500).send(`Invalid file path`)
-                }
-                const chatflowId = req.query.chatflowId as string
-                const chatId = req.query.chatId as string
-                const fileName = req.query.fileName as string
-
-                const filePath = path.join(getStoragePath(), chatflowId, chatId, fileName)
-                //raise error if file path is not absolute
-                if (!path.isAbsolute(filePath)) return res.status(500).send(`Invalid file path`)
-                //raise error if file path contains '..'
-                if (filePath.includes('..')) return res.status(500).send(`Invalid file path`)
-                //only return from the storage folder
-                if (!filePath.startsWith(getStoragePath())) return res.status(500).send(`Invalid file path`)
-
-                if (fs.existsSync(filePath)) {
-                    res.setHeader('Content-Disposition', contentDisposition(path.basename(filePath)))
-                    streamFileToUser(res, filePath)
-                } else {
-                    return res.status(404).send(`File ${fileName} not found`)
-                }
-            } catch (error) {
-                return res.status(500).send(`Invalid file path`)
-            }
-        })
 
         // ----------------------------------------
         // Configuration
